Add basic User Management support
Introduce UserService to handle user log in and allow message handling
based on the current user. If the currently logged user is the author
of the message, use a different style for it.

Also introduce simple user log in, which will be enhanced in the future
through a new component and route handling.